requirements.txt
mss>=9.0.1
Pillow>=10.0.0
pynput>=1.7.7
numpy>=1.26.0
torch>=2.2.0
